circus/web/web-requirements.txt
Mako==0.7.0
MarkupSafe==0.15
bottle==0.10.9
anyjson==0.3.1
gevent==0.13.7
gevent-socketio
gevent-websocket==0.3.6
greenlet
beaker==1.6.3
pyzmq==2.2.0.1
